+++ lib/inventree/stock.dart
@@ -220,6 +220,10 @@ class InvenTreeStockItem extends InvenTreeModel {
 
   int get status => jsondata['status'] ?? -1;
 
+  String get packaging => jsondata["packaging"] ?? "";
+
+  String get batch => jsondata["batch"] ?? "";
+
   int get partId => jsondata['part'] ?? -1;
   
   String get purchasePrice => jsondata['purchase_price'] ?? "";
@@ -242,11 +246,11 @@ class InvenTreeStockItem extends InvenTreeModel {
     }
   }
 
-  String? get updatedDateString {
+  String get updatedDateString {
     var _updated = updatedDate;
 
     if (_updated == null) {
-      return null;
+      return "";
     }
 
     final DateFormat _format = DateFormat("yyyy-MM-dd");
@@ -262,11 +266,11 @@ class InvenTreeStockItem extends InvenTreeModel {
     }
   }
 
-  String? get stocktakeDateString {
+  String get stocktakeDateString {
     var _stocktake = stocktakeDate;
 
     if (_stocktake == null) {
-      return null;
+      return "";
     }
 
     final DateFormat _format = DateFormat("yyyy-MM-dd");